requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
asynctest==0.13.0
attrs==22.2.0
azure-core==1.26.4
azure-storage-blob==12.15.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
Django==3.2.18
frozenlist==1.3.3
idna==3.4
isodate==0.6.1
multidict==6.0.4
nest-asyncio==1.5.6
pycparser==2.21
PyJWT==2.6.0
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
twilio==8.0.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
